fix(example): move providers to client component to avoid SSR serialization
createMockAuthAdapter returns an object with function properties. Passing
it directly from a Server Component to AuthProvider causes Next.js to fail
during static prerendering with "Functions cannot be passed to Client
Components". Moving provider setup into a 'use client' Providers wrapper
keeps functions client-side only.
