+++ apps/example/app/providers.tsx
@@ -1,0 +1,18 @@
+'use client';
+
+import type { ReactNode } from 'react';
+import { AppProviders } from '@jonmatum/next-shell/providers';
+import { AuthProvider } from '@jonmatum/next-shell/auth';
+import { createMockAuthAdapter } from '@jonmatum/next-shell/auth/mock';
+
+const authAdapter = createMockAuthAdapter({
+  user: { id: '1', name: 'Demo User', email: 'demo@example.com', roles: ['admin'] },
+});
+
+export function Providers({ children }: { children: ReactNode }) {
+  return (
+    <AppProviders themeProps={{ defaultTheme: 'system', enableSystem: true }}>
+      <AuthProvider adapter={authAdapter}>{children}</AuthProvider>
+    </AppProviders>
+  );
+}
